Correction of incorrect statement in README
The readme initially stated under one section that it is easier using emojinguistics to hide which languages you know (and implicit metadata that arises from such). This may be true to some extent, but the intended point was that it's a lot easier to 'change styles' using pictograms instead of words.